pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dabbler-ai-router"
version = "0.1.0"
description = "Multi-provider model routing, prompt templates, session state, and metrics for the Dabbler AI-led-workflow."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "darndestdabbler" },
]
keywords = ["ai", "router", "claude", "gemini", "openai", "session", "workflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
tests = [
    "pytest>=7.0",
    "jsonschema>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/darndestdabbler/dabbler-ai-orchestration"
Source = "https://github.com/darndestdabbler/dabbler-ai-orchestration"
Issues = "https://github.com/darndestdabbler/dabbler-ai-orchestration/issues"

[project.scripts]
close_session = "ai_router.close_session:main"
report = "ai_router.report:main"
reconciler = "ai_router.reconciler:main"
queue_status = "ai_router.queue_status:main"
heartbeat_status = "ai_router.heartbeat_status:main"
restart_role = "ai_router.restart_role:main"
backfill_session_state = "ai_router.backfill_session_state:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_router*"]

[tool.setuptools.package-data]
ai_router = [
    "router-config.yaml",
    "prompt-templates/*.md",
    "schemas/*.json",
    "docs/*.md",
]
